scraper/requirements.txt
playwright==1.49.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
